% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_generator.R
\name{key_generator}
\alias{key_generator}
\title{Obtain API key from USDA ESMIS API}
\usage{
key_generator(user.email, user.password)
}
\arguments{
\item{user.email}{a string contains the email used to login your account at https://usda.library.cornell.edu/?locale=en}

\item{user.password}{a string contains the password used to login your account at https://usda.library.cornell.edu/?locale=en}
}
\value{
A string contains your api token
}
\description{
This function helps user obtain an api token for USDA ESMIS
}
\examples{

key_generator(user.email="john.smith@example.com",user.password="password")

}
