% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_dates.R
\name{release_dates}
\alias{release_dates}
\title{Download release dates of specifict USDA announcements with report identifiers}
\usage{
release_dates(
 
    key = "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOjQ3NzJ9.J_iSIoycXn2FhgZzR53rYuyfxNoSE1F6RfhOsaagrAs",
  start_date,
  end_date,
  identifier,
  latest = F
)
}
\arguments{
\item{key}{a string contains an api key for USDA EMSI}

\item{start_date}{a string represents begining of the search period in forms of yyyy-mm-dd}

\item{end_date}{a string represents ending of the search period in form of yyyy-mm-dd}

\item{identifier}{a string represents the abbreviation of usda announcements, such as wasde}

\item{latest}{a boolean indicator. If latest is set to be TRUE, then only the latest release will be obtained. Default is latest=FALSE}
}
\value{
A vector contains release dates
}
\description{
This function uses report identifiers to download release dates of certian USDA announcements.
}
